+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 GOpinionatedTP follows [semantic versioning](https://semver.org).
 
+## unreleased
+
+- Add function to calculate default 30 second interval
+
 ## v1.0.0 (2023/09/17)
 
 Initial release